requirements.txt
streamlit==1.33.0
ipython==8.15.0
bs4==0.0.1
streamlit-chat==0.1.1
groq==0.4.2
numpy==1.24.3
google-generativeai==0.6.0